rename: mypage 파일 이동 및 파일명 수정
